Mostly fix schellingb/TinySoundFont#94
These changes seem to fix the duration of the attack mod envelope and switch to using a convex curve for it. After making this change, there seem to be some slight inaccuracies at the start of the attack mod envelope which are visible under a spectrogram when compared to e.g. FluidSynth. Committing this as-is for now because I am unsure why these occur.